refactor: Response, Request 타입에 price부분 Int가 아니라 Double 타입으로 수정 #12

+++ OpenMarket/OpenMarket/Model/Request/ProductModification.swift
@@ -12,9 +12,9 @@ struct ProductModification: Codable {
     let name: String?
     let descriptions: String?
     let thumbnailID: Int?
-    let price: Int?
+    let price: Double?
     let currency: Currency?
-    let discountedPrice: Int?
+    let discountedPrice: Double?
     let stock: Int?
     
     enum CodingKeys: String, CodingKey {
